add single point of truth for cache flush

+++ includes/functions/transient_functions.php
@@ -119,10 +119,24 @@ if (!function_exists('awm_delete_transient_all')) {
     awm_delete_transient_group($group_id);
   }
   }
+    return true;
+  }
+}
+
+if (!function_exists('ewp_flush_cache')) {
+  /**
+   * with this function we delete the transient
+   */
+  function ewp_flush_cache()
+  {
+    do_action('ewp_flush_cache_pre_action');
+    awm_delete_transient_group('awm_post_fields_transients');
+    awm_delete_transient_all();
   update_option('ewp_user_caps_version', strtotime('now'), false);
   delete_option('ewp_user_caps_version_old');
   flush_rewrite_rules();
   wp_cache_flush();
+    do_action('ewp_flush_cache_action');
     return true;
  }
 }
